% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph.R
\name{layout.forceatlas2}
\alias{layout.forceatlas2}
\title{Deprecated layout functions.}
\usage{
layout.forceatlas2(..., params = list())
}
\arguments{
\item{...}{Passed to \code{layout_with_f2}.}

\item{params}{Passed to \code{layout_with_fa2} as arguments.}
}
\description{
This is here only for compatibility with igraph's version < 0.8.0 layout API. Please use
\code{layout_with_fa2}.
}
